Start moving middlewares out in to separate files
